main: Change stat types to gint64
The JSON library will read and write the values as `gint64`, this means
that on a 32 bit system, we are incorrectly storing the stat values.

e.g. writeSectors is an unsigned long, by reading with `strtoul` and then
casting to `gint64` we avoid it becoming negative in a cast from
unsigned to signed, or dropping bits (if we were instead to use the
`int` type)

totalBytesWritten does not need to be a `double` and was also changed to a
`gint64`

Closes #54

+++ src/library/cStatComputer.cc
@@ -7,11 +7,11 @@ uint cStatComputer::getAverageWriteSize(
     return (sectorSize * pDeviceStats->writeSectors) / pDeviceStats->writeIo;
 }
 
-float cStatComputer::totalBytesWritten(uint sectorSize,
-    uintmax_t currentWriteSectors, uintmax_t previousWriteSectors,
-    float previousTotal)
+gint64 cStatComputer::totalBytesWritten(uint sectorSize,
+    gint64 currentWriteSectors, gint64 previousWriteSectors,
+    gint64 previousTotal)
 {
-    uintmax_t newTotal = previousTotal;
+    gint64 newTotal = previousTotal;
     if (currentWriteSectors < previousWriteSectors)
     {
         /* overflow occured, calculate difference:
